pyproject.toml
[tool.poetry]
name = "kg4cr"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.16"
rdflib = "^7.1.4"
openpyxl = "^3.1.5"
ipykernel = "^6.29.5"
pandas = "^2.3.0"
qlever = "^0.5.23"
bs4 = "^0.0.2"
groq = "^0.28.0"
beautifulsoup4 = "^4.13.4"
webdriver-manager = "^4.0.2"
selenium = "^4.33.0"
playwright = "^1.53.0"
undetected-chromedriver = "2.0.0"
playwright-stealth = "^2.0.0"
fake-useragent = "^2.2.0"
scrapy = "^2.13.3"
scrapy-splash = "^0.11.1"
nltk = "^3.9.1"
pyspellchecker = "^0.8.3"
symspellpy = "^6.9.0"
rapidfuzz = "^3.13.0"
requests = "^2.32.5"
google-generativeai = "^0.8.5"
perplexipy = "^1.3.1"
aiohttp = "^3.12.15"
pyvis = "^0.3.2"
python-dotenv = "^1.1.1"
tabulate = "^0.9.0"
tqdm = "^4.67.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
